% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdatum_convert.R
\name{vdatum_convert}
\alias{vdatum_convert}
\title{Convert coordinates from VDatum URL}
\usage{
vdatum_convert(df, url)
}
\arguments{
\item{df}{dataframe of input coordinates.}

\item{url}{url column as called from vdatum_url.  Must be named url}
}
\value{
a dataframe consisting of the original dataframe and cbind of the results returned from VDatum
}
\description{
Convert coordinates from VDatum URL
}
