% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdatum_url.R
\name{vdatum_url}
\alias{vdatum_url}
\title{Create a URL for NOAA VDatum API}
\usage{
vdatum_url(
  x,
  y,
  z = 0,
  region = "contiguous",
  s_h_frame = "NAD83_2011",
  s_coor = "geo",
  s_h_unit = "m",
  s_h_zone = NULL,
  s_v_frame = "NAVD88",
  s_v_unit = "m",
  s_v_elevation = "height",
  s_v_geoid = "geoid18",
  t_h_frame = "NAD83_2011",
  t_coor = "geo",
  t_h_unit = "m",
  t_h_zone = NULL,
  t_v_frame = "NAVD88",
  t_v_unit = "m",
  t_v_elevation = "height",
  t_v_geoid = "geoid18",
  epoch_in = "0.0",
  epoch_out = "0.0",
  endpoint = "https://vdatum.noaa.gov/vdatumweb/api/convert"
)
}
\arguments{
\item{x}{the column of the x coordinate (required)}

\item{y}{the column of the y coordinate (required)}

\item{z}{the colun of the z coordinate}

\item{endpoint}{}
}
\value{
a string of the url to call
}
\description{
Create a URL for NOAA VDatum API
}
